requirements.txt
uvicorn
FastAPI-SQLAlchemy==0.2.1
pydantic==1.10.10
python-dotenv==1.0.0
mysqlclient==2.2.0